moved in other file some interfaces
