+++ message_test.go
@@ -1,0 +1,91 @@
+package osc
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestMessage(t *testing.T) {
+
+	var oscAddress string = "/address"
+
+	t.Run("should append data to message", func(t *testing.T) {
+		message := NewMessage(oscAddress)
+
+		assert.Equal(t, oscAddress, message.Address)
+
+		message.Append("string argument")
+		message.Append(123456789)
+		message.Append(true)
+
+		assert.Equal(t, 3, len(message.Arguments))
+	})
+
+	t.Run("should message equal to another message", func(t *testing.T) {
+		msg1 := NewMessage(oscAddress)
+		msg2 := NewMessage(oscAddress)
+		msg1.Append(1234)
+		msg2.Append(1234)
+		msg1.Append("test string")
+		msg2.Append("test string")
+
+		assert.True(t, msg1.Equals(msg2))
+	})
+}
+
+func TestMessage_TypeTags(t *testing.T) {
+	for _, tt := range []struct {
+		desc string
+		msg  *Message
+		tags string
+		ok   bool
+	}{
+		{"addr_only", NewMessage("/"), ",", true},
+		{"nil", NewMessage("/", nil), ",N", true},
+		{"bool_true", NewMessage("/", true), ",T", true},
+		{"bool_false", NewMessage("/", false), ",F", true},
+		{"int32", NewMessage("/", int32(1)), ",i", true},
+		{"int64", NewMessage("/", int64(2)), ",h", true},
+		{"float32", NewMessage("/", float32(3.0)), ",f", true},
+		{"float64", NewMessage("/", float64(4.0)), ",d", true},
+		{"string", NewMessage("/", "5"), ",s", true},
+		{"[]byte", NewMessage("/", []byte{'6'}), ",b", true},
+		{"two_args", NewMessage("/", "123", int32(456)), ",si", true},
+	} {
+		tags := tt.msg.typeTags()
+		if got, want := tags, tt.tags; got != want {
+			t.Errorf("%s: TypeTags() = '%s', want = '%s'", tt.desc, got, want)
+		}
+	}
+}
+
+func TestMessage_String(t *testing.T) {
+	for _, tt := range []struct {
+		desc string
+		msg  *Message
+		str  string
+	}{
+		{"nil", nil, ""},
+		{"addr_only", NewMessage("/foo/bar"), "/foo/bar ,"},
+		{"one_addr", NewMessage("/foo/bar", "123"), "/foo/bar ,s 123"},
+		{"two_args", NewMessage("/foo/bar", "123", int32(456)), "/foo/bar ,si 123 456"},
+	} {
+		if got, want := tt.msg.String(), tt.str; got != want {
+			t.Errorf("%s: String() = '%s', want = '%s'", tt.desc, got, want)
+		}
+	}
+}
+
+func TestTypeTagsString(t *testing.T) {
+	msg := NewMessage("/some/address")
+	msg.Append(int32(100))
+	msg.Append(true)
+	msg.Append(false)
+
+	typeTags := msg.typeTags()
+
+	if typeTags != ",iTF" {
+		t.Errorf("Type tag string should be ',iTF' and is: %s", typeTags)
+	}
+}
