+++ osc.go
@@ -8,7 +8,6 @@ import (
 	"encoding/binary"
 	"fmt"
 	"net"
-	"reflect"
 	"regexp"
 	"strings"
 	"time"
@@ -23,29 +22,6 @@ const (
 type Packet interface {
 	MarshalBinary() (data []byte, err error)
 }
-
-// Message represents a single OSC message. An OSC message consists of an OSC
-// address pattern and zero or more arguments.
-type Message struct {
-	Address   string
-	Arguments []interface{}
-}
-
-// Verify that Messages implements the Packet interface.
-// var _ Packet = (*Message)(nil)
-
-// Bundle represents an OSC bundle. It consists of the OSC-string "#bundle"
-// followed by an OSC Time Tag, followed by zero or more OSC bundle/message
-// elements. The OSC-timetag is a 64-bit fixed point time tag. See
-// http://opensoundcontrol.org/spec-1_0 for more information.
-type Bundle struct {
-	Timetag  Timetag
-	Messages []*Message
-	Bundles  []*Bundle
-}
-
-// Verify that Bundle implements the Packet interface.
-// var _ Packet = (*Bundle)(nil)
 
 // Client enables you to send OSC packets. It sends OSC messages and bundles to
 // the given IP address and port.
@@ -62,17 +38,6 @@ type Server struct {
 	Dispatcher  Dispatcher
 	ReadTimeout time.Duration
 	close       func() error
-}
-
-// Timetag represents an OSC Time Tag.
-// An OSC Time Tag is defined as follows:
-// Time tags are represented by a 64 bit fixed point number. The first 32 bits
-// specify the number of seconds since midnight on January 1, 1900, and the
-// last 32 bits specify fractional parts of a second to a precision of about
-// 200 picoseconds. This is the representation used by Internet NTP timestamps.
-type Timetag struct {
-	timeTag uint64 // The acutal time tag
-	time    time.Time
 }
 
 // Dispatcher is an interface for an OSC message dispatcher. A dispatcher is
@@ -176,304 +141,6 @@ func (s *StandardDispatcher) Dispatch(packet Packet) {
 			}
 		}()
 	}
-}
-
-////
-// Message
-////
-
-// NewMessage returns a new Message. The address parameter is the OSC address.
-func NewMessage(addr string, args ...interface{}) *Message {
-	return &Message{Address: addr, Arguments: args}
-}
-
-// Append appends the given arguments to the arguments list.
-func (msg *Message) Append(args ...interface{}) {
-	msg.Arguments = append(msg.Arguments, args...)
-}
-
-// Equals returns true if the given OSC Message `m` is equal to the current OSC
-// Message. It checks if the OSC address and the arguments are equal. Returns
-// true if the current object and `m` are equal.
-func (msg *Message) Equals(m *Message) bool {
-	return reflect.DeepEqual(msg, m)
-}
-
-// Clear clears the OSC address and all arguments.
-func (msg *Message) Clear() {
-	msg.Address = ""
-	msg.ClearData()
-}
-
-// ClearData removes all arguments from the OSC Message.
-func (msg *Message) ClearData() {
-	msg.Arguments = msg.Arguments[len(msg.Arguments):]
-}
-
-// Match returns true, if the OSC address pattern of the OSC Message matches the given
-// address. The match is case sensitive!
-func (msg *Message) Match(addr string) bool {
-	return getRegEx(msg.Address).MatchString(addr)
-}
-
-// typeTags returns the type tag string.
-func (msg *Message) typeTags() string {
-	if len(msg.Arguments) == 0 {
-		return ","
-	}
-
-	var tags strings.Builder
-	_ = tags.WriteByte(',')
-
-	for _, m := range msg.Arguments {
-		tags.WriteByte(getTypeTag(m))
-	}
-
-	return tags.String()
-}
-
-// String implements the fmt.Stringer interface.
-func (msg *Message) String() string {
-	if msg == nil {
-		return ""
-	}
-
-	var s strings.Builder
-	tags := msg.typeTags()
-	s.WriteString(fmt.Sprintf("%s %s", msg.Address, tags))
-
-	for _, arg := range msg.Arguments {
-		switch argType := arg.(type) {
-		case bool, int32, int64, float32, float64, string:
-			s.WriteString(fmt.Sprintf(" %v", argType))
-
-		case nil:
-			s.WriteString(" Nil")
-
-		case []byte:
-			s.WriteString(fmt.Sprintf(" %s", argType))
-
-		case Timetag:
-
-			s.WriteString(fmt.Sprintf(" %d", argType.TimeTag()))
-		}
-	}
-
-	return s.String()
-}
-
-// MarshalBinary serializes the OSC message to a byte buffer. The byte buffer
-// has the following format:
-// 1. OSC Address Pattern
-// 2. OSC Type Tag String
-// 3. OSC Arguments.
-func (msg *Message) MarshalBinary() ([]byte, error) {
-	// We can start with the OSC address and add it to the buffer
-	data := new(bytes.Buffer)
-
-	_, err := writePaddedString(msg.Address, data)
-	if err != nil {
-		return nil, err
-	}
-
-	// Type tag string starts with ","
-	lenArgs := len(msg.Arguments)
-	typetags := make([]byte, lenArgs+1)
-	typetags[0] = ','
-
-	// Process the type tags and collect all arguments
-	payload := new(bytes.Buffer)
-
-	for i, arg := range msg.Arguments {
-		switch t := arg.(type) {
-		case bool:
-			if t {
-				typetags[i+1] = 'T'
-				continue
-			}
-
-			typetags[i+1] = 'F'
-
-		case nil:
-			typetags[i+1] = 'N'
-
-		case int32:
-			typetags[i+1] = 'i'
-
-			err = binary.Write(payload, binary.BigEndian, t)
-			if err != nil {
-				return nil, err
-			}
-
-		case float32:
-			typetags[i+1] = 'f'
-
-			err := binary.Write(payload, binary.BigEndian, t)
-			if err != nil {
-				return nil, err
-			}
-
-		case string:
-			typetags[i+1] = 's'
-
-			_, err = writePaddedString(t, payload)
-			if err != nil {
-				return nil, err
-			}
-
-		case []byte:
-			typetags[i+1] = 'b'
-
-			_, err = writeBlob(t, payload)
-			if err != nil {
-				return nil, err
-			}
-
-		case int64:
-			typetags[i+1] = 'h'
-
-			err = binary.Write(payload, binary.BigEndian, t)
-			if err != nil {
-				return nil, err
-			}
-
-		case float64:
-			typetags[i+1] = 'd'
-
-			err = binary.Write(payload, binary.BigEndian, t)
-			if err != nil {
-				return nil, err
-			}
-
-		case Timetag:
-			typetags[i+1] = 't'
-
-			b, err := t.MarshalBinary()
-			if err != nil {
-				return nil, err
-			}
-
-			_, err = payload.Write(b)
-			if err != nil {
-				return nil, err
-			}
-		default:
-			return nil, fmt.Errorf("unsupported type: %T", t)
-		}
-	}
-
-	// Write the type tag string to the data buffer
-	if _, err := writePaddedString(string(typetags), data); err != nil {
-		return nil, err
-	}
-
-	// Write the payload (OSC arguments) to the data buffer
-	if _, err := data.Write(payload.Bytes()); err != nil {
-		return nil, err
-	}
-
-	return data.Bytes(), nil
-}
-
-////
-// Bundle
-////
-
-// NewBundle returns an OSC Bundle. Use this function to create a new OSC
-// Bundle.
-func NewBundle(time time.Time) *Bundle {
-	return &Bundle{
-		Timetag:  *NewTimetag(time),
-		Messages: []*Message{},
-		Bundles:  []*Bundle{},
-	}
-}
-
-// Append appends an OSC bundle or OSC message to the bundle.
-func (b *Bundle) Append(pck Packet) error {
-	switch t := pck.(type) {
-	case *Bundle:
-		b.Bundles = append(b.Bundles, t)
-
-	case *Message:
-		b.Messages = append(b.Messages, t)
-
-	default:
-		return ERROR_UNSUPORTED_PACKAGE
-	}
-
-	return nil
-}
-
-// MarshalBinary serializes the OSC bundle to a byte array with the following
-// format:
-// 1. Bundle string: '#bundle'
-// 2. OSC timetag
-// 3. Length of first OSC bundle element
-// 4. First bundle element
-// 5. Length of n OSC bundle element
-// 6. n bundle element.
-func (b *Bundle) MarshalBinary() ([]byte, error) {
-	// Add the '#bundle' string
-	data := new(bytes.Buffer)
-
-	_, err := writePaddedString("#bundle", data)
-	if err != nil {
-		return nil, err
-	}
-
-	// Add the time tag
-	bd, err := b.Timetag.MarshalBinary()
-	if err != nil {
-		return nil, err
-	}
-
-	_, err = data.Write(bd)
-	if err != nil {
-		return nil, err
-	}
-
-	// Process all OSC Messages
-	for _, m := range b.Messages {
-		buf, err := m.MarshalBinary()
-		if err != nil {
-			return nil, err
-		}
-
-		// Append the length of the OSC message
-		err = binary.Write(data, binary.BigEndian, int32(len(buf)))
-		if err != nil {
-			return nil, err
-		}
-
-		// Append the OSC message
-		_, err = data.Write(buf)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Process all OSC Bundles
-	for _, b := range b.Bundles {
-		buf, err := b.MarshalBinary()
-		if err != nil {
-			return nil, err
-		}
-
-		// Write the size of the bundle
-		err = binary.Write(data, binary.BigEndian, int32(len(buf)))
-		if err != nil {
-			return nil, err
-		}
-
-		// Append the bundle
-		_, err = data.Write(buf)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return data.Bytes(), nil
 }
 
 ////
@@ -806,97 +473,6 @@ func readArguments(msg *Message, reader *bufio.Reader, start *int) error {
 	}
 
 	return nil
-}
-
-////
-// Timetag
-////
-
-// NewTimetag returns a new OSC time tag object.
-func NewTimetag(timeStamp time.Time) *Timetag {
-	return &Timetag{
-		time:    timeStamp,
-		timeTag: timeToTimetag(timeStamp),
-	}
-}
-
-// NewTimetagFromTimetag creates a new Timetag from the given `timetag`.
-func NewTimetagFromTimetag(timetag uint64) *Timetag {
-	time := timetagToTime(timetag)
-	return NewTimetag(time)
-}
-
-// Time returns the time.
-func (t *Timetag) Time() time.Time {
-	return t.time
-}
-
-// FractionalSecond returns the last 32 bits of the OSC time tag. Specifies the
-// fractional part of a second.
-func (t *Timetag) FractionalSecond() uint32 {
-	return uint32(t.timeTag << 32)
-}
-
-// SecondsSinceEpoch returns the first 32 bits (the number of seconds since the
-// midnight 1900) from the OSC time tag.
-func (t *Timetag) SecondsSinceEpoch() uint32 {
-	return uint32(t.timeTag >> 32)
-}
-
-// TimeTag returns the time tag value.
-func (t *Timetag) TimeTag() uint64 {
-	return t.timeTag
-}
-
-// MarshalBinary converts the OSC time tag to a byte array.
-func (t *Timetag) MarshalBinary() ([]byte, error) {
-	data := new(bytes.Buffer)
-	err := binary.Write(data, binary.BigEndian, t.timeTag)
-	return data.Bytes(), err
-}
-
-// SetTime sets the value of the OSC time tag.
-func (t *Timetag) SetTime(time time.Time) {
-	t.time = time
-	t.timeTag = timeToTimetag(time)
-}
-
-// ExpiresIn calculates the number of seconds until the current time is the
-// same as the value of the time tag. It returns zero if the value of the
-// time tag is in the past.
-func (t *Timetag) ExpiresIn() time.Duration {
-	if t.timeTag <= 1 {
-		return 0
-	}
-
-	tt := timetagToTime(t.timeTag)
-	seconds := time.Until(tt)
-
-	if seconds <= 0 {
-		return 0
-	}
-
-	return seconds
-}
-
-// timeToTimetag converts the given time to an OSC time tag.
-//
-// An OSC time tag is defined as follows:
-// Time tags are represented by a 64 bit fixed point number. The first 32 bits
-// specify the number of seconds since midnight on January 1, 1900, and the
-// last 32 bits specify fractional parts of a second to a precision of about
-// 200 picoseconds. This is the representation used by Internet NTP timestamps.
-//
-// The time tag value consisting of 63 zero bits followed by a one in the least
-// significant bit is a special case meaning "immediately.".
-func timeToTimetag(time time.Time) (timetag uint64) {
-	timetag = uint64((secondsFrom1900To1970 + time.Unix()) << 32)
-	return (timetag + uint64(uint32(time.Nanosecond())))
-}
-
-// timetagToTime converts the given timetag to a time object.
-func timetagToTime(timetag uint64) (t time.Time) {
-	return time.Unix(int64((timetag>>32)-secondsFrom1900To1970), int64(timetag&0xffffffff))
 }
 
 ////
